requirements.txt
# requirements.txt

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# CORS
starlette==0.27.0

fastapi_mail

psycopg2

sqlalchemy_utils

motor

python-multipart